Move all reservations concurrency logic to reservations service
