+++ src/backend/Exceptions/QuantityTooLargeException.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace backend.Exceptions
+{
+    public class QuantityTooLargeException : Exception
+    {
+        public QuantityTooLargeException() {}
+    }
+}
